QCamera: Interface cleanup
- All size member variables changed to size_t.
  Others changed to named size types, e.g
  int changed to int32_t, etc.
- Moved macro CAM_DUMP_TO_FILE to
  'mm_camera_interface.h'.
- Removed irrelevant macro FLIP_V_H

CRs-Fixed: 641033

+++ QCamera2/HAL/QCamera2HWI.cpp
@@ -1156,7 +1156,7 @@ int QCamera2HardwareInterface::openCamera()
     m_max_pic_height = 0;
     char value[PROPERTY_VALUE_MAX];
     int enable_4k2k;
-    int i;
+    size_t i;
 
     if (mCameraHandle) {
         ALOGE("Failure: Camera already opened");
@@ -5923,7 +5923,7 @@ int QCamera2HardwareInterface::calcThermalLevel(
             // Set lowest min FPS for now
             adjustedRange.min_fps = minFPS/1000.0f;
             adjustedRange.max_fps = minFPS/1000.0f;
-            for ( int i = 0 ; i < gCamCapability[mCameraId]->fps_ranges_tbl_cnt ; i++ ) {
+            for ( size_t i = 0 ; i < gCamCapability[mCameraId]->fps_ranges_tbl_cnt ; i++ ) {
                 if ( gCamCapability[mCameraId]->fps_ranges_tbl[i].min_fps < adjustedRange.min_fps ) {
                     adjustedRange.min_fps = gCamCapability[mCameraId]->fps_ranges_tbl[i].min_fps;
                     adjustedRange.max_fps = adjustedRange.min_fps;
